+++ config/messages.py
@@ -481,10 +481,3 @@ class Messages(metaclass=Formatable):
     timeout_wars_user = "Uživatel {user} byl umlčen na {time:.0f} minut."
     timeout_wars_message_delete = "Uživatel {user} byl přistižen při mazání zpráv. Byl proto umlčen na {time:.0f} minut."
     timeout_wars_user_immunity = "Uživatel {user} má ještě imunitu na {time:.0f} sekund."
-
-    # MODERATION
-    slowmode_time = "Délka prodlevy v sekundách (vyber v autocomplete)"
-    slowmode_set_brief = "Nastaví slowmode v aktuálním kanálu"
-    slowmode_set_success = "Slowmode v kanálu {channel} nastaven na {delay} sekund."
-    slowmode_remove_brief = "Vypne slowmode v aktuálním kanálu"
-    slowmode_remove_success = "Slowmode v kanálu {channel} úspěšně odstraněn."
